requirements-pip.txt
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037492802/work
async-generator==1.10
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
certifi==2021.5.30
cffi==1.14.0
charset-normalizer==2.0.12
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dill @ file:///tmp/build/80754af9/dill_1623919422540/work
docopt==0.6.2
entrypoints==0.3
graphviz==0.19.1
idna==3.4
importlib-resources==5.4.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi==0.17.0
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
jsonschema==3.0.2
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1633420100582/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282412546/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528150516/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///tmp/build/80754af9/notebook_1629197205351/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1625649052827/work
pipreqs==0.4.11
plotly @ file:///opt/conda/conda-bld/plotly_1646671701182/work
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyzmq @ file:///tmp/build/80754af9/pyzmq_1628268400816/work
requests==2.27.1
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
statsmodels @ file:///tmp/build/80754af9/statsmodels_1614023745203/work
tenacity @ file:///tmp/build/80754af9/tenacity_1632756926792/work
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
torch==1.10.2
torchvision==0.4.2
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1632746497744/work
typing_extensions==4.1.1
urllib3==1.26.13
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==3.6.0
